@@ -1,79 +1,85 @@
 package com.emedinaa.kotlincoroutines
 
 import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.util.Log
+import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.lifecycleScope
 import com.emedinaa.kotlincoroutines.data.RemoteDataSource
 import com.emedinaa.kotlincoroutines.data.Repository
-import kotlinx.android.synthetic.main.activity_main.*
+import com.emedinaa.kotlincoroutines.databinding.ActivityMainBinding
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.async
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 
+/**
+ * @author Eduardo Medina
+ */
 class MainActivity : AppCompatActivity() {
 
-    private var reviewList:List<Review> = emptyList()
+    private lateinit var binding: ActivityMainBinding
 
-    private val adapter:MainAdapter by lazy {
-        MainAdapter(emptyList()){
+    private val adapter: MainAdapter by lazy {
+        MainAdapter(emptyList()) {
             goToCourse(it)
         }
     }
-    private val repository:Repository by  lazy {
+
+    private val repository: Repository by lazy {
         Repository(RemoteDataSource())
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-        recyclerView.adapter = adapter
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+        binding.recyclerView.adapter = adapter
 
         fetchCourses()
         //fetchData()
     }
 
-    private fun goToCourse(course: Course){
-        val intent = Intent(this,CourseActivity::class.java)
+    private fun goToCourse(course: Course) {
+        val intent = Intent(this, CourseActivity::class.java)
         intent.putExtras(Bundle().apply {
-            putParcelable("COURSE",course)
+            putParcelable("COURSE", course)
         })
         startActivity(intent)
     }
 
-    private fun fetchCourses(){
+    private fun fetchCourses() {
         lifecycleScope.launch {
-            val result = withContext(Dispatchers.IO){
+            val result = withContext(Dispatchers.IO) {
+                repository.fetchCourses()
+            }
+            Log.v("CONSOLE","result $result")
+            adapter.update(result)
+        }
+    }
+
+    private fun fetchReviews() {
+        lifecycleScope.launch {
+            val result = withContext(Dispatchers.IO) {
                 repository.fetchCourses()
             }
             adapter.update(result)
         }
     }
 
-    private fun fetchReviews(){
+    private fun fetchData() {
         lifecycleScope.launch {
-            val result = withContext(Dispatchers.IO){
+            val courseDeferred = async(Dispatchers.IO) {
                 repository.fetchCourses()
             }
-            adapter.update(result)
-        }
-    }
-
-    private fun fetchData(){
-        lifecycleScope.launch {
-            val courseDeferred = async(Dispatchers.IO){
-                repository.fetchCourses()
-            }
-            val reviewDeferred = async(Dispatchers.IO){
+            val reviewDeferred = async(Dispatchers.IO) {
                 repository.fetchReviews()
             }
 
             val courseResult = courseDeferred.await()
-            val reviewResult = reviewDeferred.await()
+            reviewDeferred.await()
 
             adapter.update(courseResult)
-            reviewList = reviewResult
         }
     }
 
