@@ -3,10 +3,14 @@ package com.emedinaa.kotlincoroutines
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
-import kotlinx.android.synthetic.main.row_review.view.*
 
-class ReviewAdapter(private var reviews:List<Review>):RecyclerView.Adapter<ReviewAdapter.ReviewHolder>(){
+/**
+ * @author Eduardo Medina
+ */
+class ReviewAdapter(private var reviews: List<Review>) :
+    RecyclerView.Adapter<ReviewAdapter.ReviewHolder>() {
 
     override fun onCreateViewHolder(parent: ViewGroup, p1: Int): ReviewHolder {
         val view = LayoutInflater.from(parent.context)
@@ -18,19 +22,20 @@ class ReviewAdapter(private var reviews:List<Review>):RecyclerView.Adapter<Revie
         vh.bind(reviews[position])
     }
 
-    override fun getItemCount(): Int {
-        return reviews.size
-    }
+    override fun getItemCount(): Int = reviews.size
 
-    fun update(data:List<Review>){
-        reviews= data
+    fun update(data: List<Review>) {
+        reviews = data
         notifyDataSetChanged()
     }
 
-    inner class ReviewHolder(private val view: View) : RecyclerView.ViewHolder(view){
-        fun bind(entity:Review){
-            view.textViewAuthor.text = entity.author
-            view.textViewDesc.text = entity.comment
+    inner class ReviewHolder(view: View) : RecyclerView.ViewHolder(view) {
+        private val textViewAuthor: TextView = view.findViewById(R.id.textViewAuthor)
+        private val textViewDesc: TextView = view.findViewById(R.id.textViewDesc)
+
+        fun bind(entity: Review) {
+            textViewAuthor.text = entity.author
+            textViewDesc.text = entity.comment
         }
     }
 }
